requirements.txt
# Web Framework
flask>=2.0.0
flask_session>=0.4.0
werkzeug>=2.0.0

# Data Processing
pandas>=1.3.0
numpy>=1.20.0
chardet>=4.0.0  # For file encoding detection

# Machine Learning
scikit-learn>=1.0.0
joblib>=1.0.0
xgboost>=1.5.0
lightgbm
catboost

statsmodels>=0.13.0
imbalanced-learn>=0.9.0  # For handling imbalanced datasets

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# API Integration
requests>=2.25.0
google-generativeai

# Authentication
flask_sqlalchemy>=2.5.0
flask_login>=0.5.0

# Optional ML Extensions
pycaret>=2.3.0  # Automated ML
auto-sklearn>=0.14.0  # Automated ML
optuna>=2.10.0  # Hyperparameter optimization
shap>=0.40.0  # Model explainability
eli5>=0.11.0  # Model explainability
lime>=0.2.0  # Model explainability
yellowbrick>=1.4  # ML visualization

# Optional Google Generative AI
google-generativeai>=0.1.0

# File handling
openpyxl>=3.0.0  # For Excel file support
python-docx>=0.8.0  # For DOCX file support

# Logging and debugging
traceback>=0.4.0
logging>=0.5.0python-dotenv==0.19.1
